request-hander basics work
You should be able to run request-sharder.py with an on disk sqlite3
database and fire requests off to it. With curl, you can try
```
curl --header 'REMOTE_USER: foo' 127.0.0.1:8888
```
A user called `foo` should be sharded and placed into the database. If
the request is repeated the log will indicate the record is already
found and return the hub name.

There are some major caveats though, this doesn't work using a memory
based sqlite database (probably because of my crappy session handling),
and the sqlite database must be deleted after each run (initialization
doesn't understand how to import existing records yet and will try to
reinitialize the dummy entries).

Signed-off-by: Ian Allison <iana@pims.math.ca>